Fix configuration value loading
git-config appears to lowercase keys before returning them, which
meant sync.excludePaths and friends weren't getting found. Look them
up lowercased, instead.

I'm not sure if this is a change in behavior on git's side (it does
this as of 2.20.1, at least), so to make sure, lowercase it outselves
at the ingestion point.